skeleton

+++ src/vollseg/utils.py
@@ -4390,7 +4390,7 @@ def CellPoseWater(membrane_image, sized_smart_seeds, mask, decay_rate = 1.0):
     for decay_map in decay_maps:
         markers *= decay_map[:, np.newaxis, np.newaxis]
     
-
+    markers = markers.astype(np.uint16)
     thresh = threshold_otsu(membrane_image)
     binary_image = membrane_image > thresh
     thick_binary_image = binary_image.copy()
